% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eur.R
\name{eur}
\alias{eur}
\title{Formats a number to a string using the Euro sign, rounds up to 2 digits and separates with a comma}
\usage{
eur(amt)
}
\arguments{
\item{amt}{amount to be formatted}
}
\value{
string
}
\description{
Formats a number to a string using the Euro sign, rounds up to 2 digits and separates with a comma
}
\examples{
eur(1)
eur(25275.156)
eur(-31525)
}
