% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurusd.R
\name{eurusd}
\alias{eurusd}
\title{Looks up the most recent USD to EUR exchange rate via an API and returns the rate as a number}
\usage{
eurusd()
}
\value{
number
}
\description{
Looks up the most recent USD to EUR exchange rate via an API and returns the rate as a number
}
\examples{
eurusd()
}
