% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneur.R
\name{uneur}
\alias{uneur}
\title{Transforms a string into a number}
\usage{
uneur(str)
}
\arguments{
\item{eur}{value to be transformed}
}
\value{
integer
}
\description{
Transforms a string into a number
}
\examples{
uneur('€1')
uneur('€25,275.156')
uneur('€-31,525')
}
