% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_usd_to_eur.R
\name{convert_usd_to_eur}
\alias{convert_usd_to_eur}
\title{Looks up the most recent USD to EUR exchange rate and computes the provided USD amount in EUR}
\usage{
convert_usd_to_eur(usd)
}
\arguments{
\item{usd}{amount in USD to be converted to EUR}
}
\value{
number
}
\description{
Looks up the most recent USD to EUR exchange rate and computes the provided USD amount in EUR
}
\examples{
convert_usd_to_eur(10)
convert_usd_to_eur(25)
}
